ssss/requirements.txt
flask
python-dotenv
groq
pdf2image
werkzeug
gunicorn
flask-cors
Pillow